✨ Add search by partial names
Right now we are storing ingredients as a string with all the information "1 cup of flour". Go to the readme to understand why we are doing this way. We can expect users not to give the quantity of what they have but just the food that they have: "flour". Also, users could try "cocoa" when the recipe has "cocoa powder"

This changes the search for a partial one. It could have some problems scaling and using other approaches like Postgres's full-text search could help.

- Update search for a partial search chaining multiple conditions
- Add new test for partial matches on ingredients
- Update tests to use just the food as search terms